refactor(encoder): simplify code and run test from python
